requirements.txt
fastapi
uvicorn[standard]
openai
redis
sqlalchemy
asyncpg
psycopg2-binary
pydantic
python-dotenv
typing-extensions
orjson


cachetools

httpx
requests

loguru
sentry-sdk


databases[postgresql]
email-validator

black
ruff
isort
mypy
pre-commit
types-redis
types-requests
types-python-dateutil

pytest
pytest-asyncio
pytest-cov

click
python-multipart

fastapi
uvicorn
pyjwt
httpx
python-dotenv
jsonschema==4.21.0

fastapi
uvicorn[standard]
pydantic
jsonschema
httpx